14-th sprint, first review

+++ controllers/cards.js
@@ -1,14 +1,14 @@
 const Card = require('../models/card');
-const { ERROR_CODE, NOT_FOUND, UNDEFINED_ERROR } = require('../errors/error-const');
 const NotFoundError = require('../errors/not-found');
+const BadRequestError = require('../errors/unathorized');
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .then((card) => res.status(200).send(card))
-    .catch(() => res.status(UNDEFINED_ERROR).send({ message: 'Ошибка по умолчанию' }));
+    .catch(next);
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   const owner = req.user._id;
 
@@ -16,29 +16,46 @@ module.exports.createCard = (req, res) => {
     .then((card) => res.status(200).send(card))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        res.status(ERROR_CODE).send({ message: 'Переданы некорректные данные при создании карточки' });
+        next(new BadRequestError('Переданы некорректные данные при создании карточки'));
       } else {
-        res.status(UNDEFINED_ERROR).send({ message: 'Ошибка по умолчанию' });
+        next(err);
       }
     });
 };
 
-module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
+module.exports.deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
     .orFail(() => new NotFoundError('NotFoundError'))
-    .then((card) => res.status(200).send(card))
-    .catch((err) => {
-      if (err.message === 'NotFoundError') {
-        res.status(NOT_FOUND).send({ message: 'Карточка с указанным _id не найдена' });
-      } if (err.name === 'CastError') {
-        res.status(ERROR_CODE).send({ message: 'Невалидный id' });
+    .then((cardOwner) => {
+      if (cardOwner.owner.equals(req.user._id)) {
+        Card.findByIdAndRemove(req.params.cardId)
+          .orFail(() => new NotFoundError('NotFoundError'))
+          .then((card) => res.status(200).send(card))
+          .catch((err) => {
+            if (err.message === 'NotFoundError') {
+              next(new NotFoundError('Карточка с указанным _id не найдена'));
+            } if (err.name === 'CastError') {
+              next(new BadRequestError('Невалидный id'));
+            } else {
+              next(err);
+            }
+          });
       } else {
-        res.status(NOT_FOUND).send({ message: 'Карточка с указанным _id не найдена' });
+        next(new BadRequestError('Удаление карточек других пользователей запрещено'));
+      }
+    })
+    .catch((err) => {
+      if (err.name === 'CastError') {
+        next(new BadRequestError('Невалидный id'));
+      } else if (err.message === 'NotFoundError') {
+        next(new NotFoundError('Карточка с указанным _id не найдена'));
+      } else {
+        next(err);
       }
     });
 };
 
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } }, // добавить _id в массив, если его там нет
@@ -48,18 +65,18 @@ module.exports.likeCard = (req, res) => {
     .then((card) => res.status(200).send(card))
     .catch((err) => {
       if (err.message === 'NotFoundError') {
-        res.status(NOT_FOUND).send({ message: 'Передан несуществующий _id карточки' });
+        next(new NotFoundError('Передан несуществующий _id карточки'));
       } else if (err.name === 'CastError') {
-        res.status(ERROR_CODE).send({ message: 'Невалидный id' });
+        next(new BadRequestError('Невалидный id'));
       } else if (err.name === 'ValidationError') {
-        res.status(ERROR_CODE).send({ message: 'Переданы некорректные данные для постановки/снятии лайка' });
+        next(new BadRequestError('Переданы некорректные данные для постановки лайка'));
       } else {
-        res.status(UNDEFINED_ERROR).send({ message: 'Ошибка по умолчанию' });
+        next(err);
       }
     });
 };
 
-module.exports.dislikeCard = (req, res) => {
+module.exports.dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } }, // убрать _id из массива
@@ -69,13 +86,13 @@ module.exports.dislikeCard = (req, res) => {
     .then((card) => res.status(200).send(card))
     .catch((err) => {
       if (err.message === 'NotFoundError') {
-        res.status(NOT_FOUND).send({ message: 'Передан несуществующий _id карточки' });
+        next(new NotFoundError('Передан несуществующий _id карточки'));
       } else if (err.name === 'CastError') {
-        res.status(ERROR_CODE).send({ message: 'Невалидный id' });
+        next(new BadRequestError('Невалидный id'));
       } else if (err.name === 'ValidationError') {
-        res.status(ERROR_CODE).send({ message: 'Переданы некорректные данные для постановки/снятии лайка' });
+        next(new BadRequestError('Переданы некорректные данные для снятии лайка'));
       } else {
-        res.status(UNDEFINED_ERROR).send({ message: 'Ошибка по умолчанию' });
+        next(err);
       }
     });
 };
